Fixes (#996)
* Fix "BigNumber.toString does not accept any parameters" warning

* Bump TypeScript version to 5.2.2

* Bump denedencies, 'eslint' version, and run 'eslint --fix'

* Fix lint

+++ packages/contracts/src/test/TestDecimalsToken.sol
@@ -9,7 +9,7 @@ contract TestDecimalsToken is ERC20 {
         mint(100 ether);
     }
 
-    function mint(uint amount) public {
+    function mint(uint256 amount) public {
         _mint(msg.sender, amount);
     }
